+++ packages/active-campaign-client/src/__tests__/handlers/deleteContact.test.ts
@@ -2,7 +2,7 @@ import { handler } from '../../handlers/deleteContact';
 import { SQSEvent } from 'aws-lambda';
 
 // remove .skip to run the test, be aware it does a real API call so it will create a contact in the active campaign account
-describe.skip('deleteContact handler', () => {
+describe('deleteContact handler', () => {
   it('should delete a contact successfully', async () => {
     const event: SQSEvent = {
       Records: [
@@ -11,7 +11,7 @@ describe.skip('deleteContact handler', () => {
           receiptHandle: '1',
           body: JSON.stringify({
             // Replace this with the existing email of the contact you want to delete, otherwise the test will fail
-            username: `test@example1731961388029e.com`,
+            cognitoId: `466e0280-9061-7007-c3e0-beb6be672f68`,
           }),
           attributes: {
             ApproximateReceiveCount: '1',
