+++ packages/active-campaign-client/src/handlers/deleteContact.ts
@@ -1,5 +1,5 @@
 import { APIGatewayProxyResult, SQSEvent } from 'aws-lambda';
-import { acClient } from '../activeCampaignClient';
+import { acClient } from '../utils/activeCampaignClient';
 
 export async function handler(event: {
   readonly Records: SQSEvent['Records'];
@@ -7,9 +7,9 @@ export async function handler(event: {
   try {
     const firstMessage = event.Records[0] ?? { body: '{}' };
     // Parse request body
-    const { username } = JSON.parse(firstMessage.body);
+    const { cognitoId } = JSON.parse(firstMessage.body);
 
-    const contactId = await acClient.getContactByEmail(username);
+    const contactId = await acClient.getContactByCognitoId(cognitoId);
 
     if (!contactId) {
       return {
