+++ packages/active-campaign-client/src/handlers/updateContact.ts
@@ -1,5 +1,5 @@
 import { APIGatewayProxyResult, SQSEvent } from 'aws-lambda';
-import { acClient } from '../activeCampaignClient';
+import { acClient } from '../utils/activeCampaignClient';
 import { ContactPayload } from '../types/contactPayload';
 
 export async function handler(event: {
@@ -17,7 +17,7 @@ export async function handler(event: {
       };
     }
 
-    const contactId = await acClient.getContactByEmail(userData.username);
+    const contactId = await acClient.getContactByCognitoId(userData.cognitoId);
 
     if (!contactId) {
       return {
