+++ packages/active-campaign-client/src/index.ts
@@ -1,0 +1,3 @@
+import { handler } from './handlers/addContact';
+
+export { handler };
