+++ packages/active-campaign-client/src/types/webinarPayload.ts
@@ -1,0 +1,5 @@
+export type WebinarPayload = {
+  readonly slug: string;
+  readonly title: string;
+  readonly subscribeCtaLabel: string;
+};
